Refactored io into a sub-package
